+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+### 1.8.0 (2016-5-17)
+
+##### New Features
+
+* **collections:** add support for bookshelf collections ([4cae0206](https://github.com/lob/hapi-bookshelf-models/commit/4cae0206a35679f8b176d6d71f958b5949ab6611))
+
+#### 1.7.1 (2016-2-11)
+
+##### Bug Fixes
+
+* **paths:** fixed error thrown by a configuration with a relative path to models ([efadcdd4](https://github.com/lob/hapi-bookshelf-models/commit/efadcdd4b6f36fa197b27577eeeb205cef12c710))
+
+### 1.7.0 (2016-2-2)
+
+##### New Features
+
+* **paths:** added glob to support wildcard model folder structure ([68fc0f5a](https://github.com/lob/hapi-bookshelf-models/commit/68fc0f5a8e888204145d9ba4f1764b5efb7bdb5b))
+
+##### Chores
+
+* **release:** add release scripts with changelog generation ([63bbf46f](https://github.com/lob/hapi-bookshelf-models/commit/63bbf46f))
+
+## 1.6.0 (2015-12-23)
+
+- Update bookshelf to 0.9.1
+
+## 1.5.0 (2015-09-29)
+
+- Allow Hapi 9.x and 10.x
+
 ## 1.4.0 (2015-09-10)
 
 - Ignore hidden files in models directory